+++ processes/gutenberg.py
@@ -20,10 +20,10 @@ class GutenbergProcess(CoreProcess):
     }
 
     def __init__(self, *args):
-        super(GutenbergProcess, self).__init__(*args[:3])
+        super(GutenbergProcess, self).__init__(*args[:4])
 
-        self.ingestOffset = int(args[4]) or 0
-        self.ingestLimit = (int(args[3]) + self.ingestOffset) or 5000
+        self.ingestOffset = int(args[5]) or 0
+        self.ingestLimit = (int(args[4]) + self.ingestOffset) or 5000
 
         # Connect to database
         self.generateEngine()