+++ tests/helper.py
@@ -38,7 +38,8 @@ class TestHelpers:
         'GITHUB_API_KEY': 'test_github_key',
         'GITHUB_API_ROOT': 'test_github_url',
         'BARDO_CCE_API': 'test_cce_url',
-        'MUSE_MARC_URL': 'test_muse_url'
+        'MUSE_MARC_URL': 'test_muse_url',
+        'DOAB_OAI_URL': 'test_doab_url'
     }
 
     @classmethod